feat: use serde_json::Value, Discord and Dotenvy
Instead of using the full structs to exactly mimic what Cloudflare does,
serde_json::Value is used, to only work with the values needed for the A
record.
When an error is received from Cloudflare, display the error.
When the outside IP has changed, send a message via a Discord webhook.
Use Dotenvy to setup environment variables first. Then, Clap is used for
regular command-line argument processing.